Add example of disabling mouse hover with for loop
I added an example of disabling the mouse hover with a loop rather than individual calls, as well as showing what you might use if there were multiple VBoxContainers, aka "menus", under on larger parent scene. You could do the same method with enabling the mouse hover as well on the other part of the code. I just left that alone to keep the original example for part of the code still like in the video.